+++ crates/fuel-core/src/schema/tx/types.rs
@@ -21,6 +21,7 @@ use crate::{
         block::Block,
         scalars::{
             AssetId,
+            BlobId,
             Bytes32,
             ContractId,
             HexString,
@@ -53,6 +54,7 @@ use fuel_core_types::{
         field::{
             BytecodeRoot,
             BytecodeWitnessIndex,
+            ChargeableBody,
             InputContract,
             Inputs,
             Maturity,
@@ -86,8 +88,10 @@ use fuel_core_types::{
             TransactionExecutionResult,
             TransactionExecutionStatus,
         },
-        txpool,
-        txpool::TransactionStatus as TxStatus,
+        txpool::{
+            self,
+            TransactionStatus as TxStatus,
+        },
     },
     tai64::Tai64,
 };
@@ -426,6 +430,11 @@ impl Transaction {
                     .map(|c| AssetId(*c))
                     .collect(),
             ),
+            fuel_tx::Transaction::Blob(tx) => Some(
+                tx.input_asset_ids(base_asset_id)
+                    .map(|c| AssetId(*c))
+                    .collect(),
+            ),
         }
     }
 
@@ -446,6 +455,9 @@ impl Transaction {
             fuel_tx::Transaction::Upload(tx) => {
                 Some(input_contracts(tx).map(|v| (*v).into()).collect())
             }
+            fuel_tx::Transaction::Blob(tx) => {
+                Some(input_contracts(tx).map(|v| (*v).into()).collect())
+            }
         }
     }
 
@@ -454,7 +466,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Mint(mint) => Some(mint.input_contract().into()),
         }
     }
@@ -466,6 +479,7 @@ impl Transaction {
             fuel_tx::Transaction::Mint(_) => None,
             fuel_tx::Transaction::Upgrade(tx) => Some((*tx.policies()).into()),
             fuel_tx::Transaction::Upload(tx) => Some((*tx.policies()).into()),
+            fuel_tx::Transaction::Blob(tx) => Some((*tx.policies()).into()),
         }
     }
 
@@ -477,7 +491,8 @@ impl Transaction {
             fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
         }
     }
 
@@ -488,6 +503,7 @@ impl Transaction {
             fuel_tx::Transaction::Mint(_) => None,
             fuel_tx::Transaction::Upgrade(tx) => Some(tx.maturity().into()),
             fuel_tx::Transaction::Upload(tx) => Some(tx.maturity().into()),
+            fuel_tx::Transaction::Blob(tx) => Some(tx.maturity().into()),
         }
     }
 
@@ -496,7 +512,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Mint(mint) => Some((*mint.mint_amount()).into()),
         }
     }
@@ -506,7 +523,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Mint(mint) => Some((*mint.mint_asset_id()).into()),
         }
     }
@@ -516,7 +534,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Mint(mint) => Some((*mint.gas_price()).into()),
         }
     }
@@ -527,7 +546,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Mint(mint) => Some((*mint.tx_pointer()).into()),
         }
     }
@@ -552,6 +572,10 @@ impl Transaction {
         self.0.is_upload()
     }
 
+    async fn is_blob(&self) -> bool {
+        self.0.is_blob()
+    }
+
     async fn inputs(&self) -> Option<Vec<Input>> {
         match &self.0 {
             fuel_tx::Transaction::Script(tx) => {
@@ -565,6 +589,9 @@ impl Transaction {
                 Some(tx.inputs().iter().map(Into::into).collect())
             }
             fuel_tx::Transaction::Upload(tx) => {
+                Some(tx.inputs().iter().map(Into::into).collect())
+            }
+            fuel_tx::Transaction::Blob(tx) => {
                 Some(tx.inputs().iter().map(Into::into).collect())
             }
         }
@@ -585,6 +612,9 @@ impl Transaction {
             fuel_tx::Transaction::Upload(tx) => {
                 tx.outputs().iter().map(Into::into).collect()
             }
+            fuel_tx::Transaction::Blob(tx) => {
+                tx.outputs().iter().map(Into::into).collect()
+            }
         }
     }
 
@@ -593,7 +623,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Mint(mint) => Some(mint.output_contract().into()),
         }
     }
@@ -625,6 +656,12 @@ impl Transaction {
                     .map(|w| HexString(w.clone().into_inner()))
                     .collect(),
             ),
+            fuel_tx::Transaction::Blob(tx) => Some(
+                tx.witnesses()
+                    .iter()
+                    .map(|w| HexString(w.clone().into_inner()))
+                    .collect(),
+            ),
         }
     }
 
@@ -636,7 +673,8 @@ impl Transaction {
             fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
         }
     }
 
@@ -659,7 +697,8 @@ impl Transaction {
             fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
         }
     }
 
@@ -671,7 +710,8 @@ impl Transaction {
             fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
             | fuel_tx::Transaction::Upgrade(_)
-            | fuel_tx::Transaction::Upload(_) => None,
+            | fuel_tx::Transaction::Upload(_)
+            | fuel_tx::Transaction::Blob(_) => None,
         }
     }
 
@@ -686,6 +726,14 @@ impl Transaction {
             fuel_tx::Transaction::Upload(tx) => {
                 Some((*tx.bytecode_witness_index()).into())
             }
+            fuel_tx::Transaction::Blob(tx) => Some((*tx.bytecode_witness_index()).into()),
+        }
+    }
+
+    async fn blob_id(&self) -> Option<BlobId> {
+        match &self.0 {
+            fuel_tx::Transaction::Blob(blob) => Some(blob.body().id.into()),
+            _ => None,
         }
     }
 
@@ -696,6 +744,7 @@ impl Transaction {
             fuel_tx::Transaction::Mint(_) => None,
             fuel_tx::Transaction::Upgrade(_) => None,
             fuel_tx::Transaction::Upload(_) => None,
+            fuel_tx::Transaction::Blob(_) => None,
         }
     }
 
@@ -721,6 +770,7 @@ impl Transaction {
             fuel_tx::Transaction::Mint(_) => None,
             fuel_tx::Transaction::Upgrade(_) => None,
             fuel_tx::Transaction::Upload(_) => None,
+            fuel_tx::Transaction::Blob(_) => None,
         }
     }
 
@@ -729,7 +779,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
-            | fuel_tx::Transaction::Upgrade(_) => None,
+            | fuel_tx::Transaction::Upgrade(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Upload(tx) => Some((*tx.bytecode_root()).into()),
         }
     }
@@ -739,7 +790,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
-            | fuel_tx::Transaction::Upgrade(_) => None,
+            | fuel_tx::Transaction::Upgrade(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Upload(tx) => Some((*tx.subsection_index()).into()),
         }
     }
@@ -749,7 +801,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
-            | fuel_tx::Transaction::Upgrade(_) => None,
+            | fuel_tx::Transaction::Upgrade(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Upload(tx) => Some((*tx.subsections_number()).into()),
         }
     }
@@ -759,7 +812,8 @@ impl Transaction {
             fuel_tx::Transaction::Script(_)
             | fuel_tx::Transaction::Create(_)
             | fuel_tx::Transaction::Mint(_)
-            | fuel_tx::Transaction::Upgrade(_) => None,
+            | fuel_tx::Transaction::Upgrade(_)
+            | fuel_tx::Transaction::Blob(_) => None,
             fuel_tx::Transaction::Upload(tx) => {
                 Some(tx.proof_set().iter().map(|proof| (*proof).into()).collect())
             }
@@ -773,6 +827,7 @@ impl Transaction {
             | fuel_tx::Transaction::Mint(_) => None,
             fuel_tx::Transaction::Upgrade(tx) => Some((*tx.upgrade_purpose()).into()),
             fuel_tx::Transaction::Upload(_) => None,
+            fuel_tx::Transaction::Blob(_) => None,
         }
     }