+++ crates/fuel-core/src/service/sub_services.rs
@@ -1,4 +1,5 @@
 #![allow(clippy::let_unit_value)]
+
 use super::{
     adapters::P2PAdapter,
     genesis::create_genesis_block,
@@ -32,22 +33,30 @@ use crate::{
         SubServices,
     },
 };
+#[allow(unused_imports)]
 use fuel_core_gas_price_service::fuel_gas_price_updater::{
     fuel_core_storage_adapter::FuelL2BlockSource,
     Algorithm,
+    AlgorithmV0,
     FuelGasPriceUpdater,
     UpdaterMetadata,
     V0Metadata,
 };
 use fuel_core_poa::Trigger;
+use fuel_core_services::{
+    RunnableService,
+    ServiceRunner,
+};
 use fuel_core_storage::{
-    structured_storage::StructuredStorage,
+    self,
     transactional::AtomicView,
 };
 #[cfg(feature = "relayer")]
 use fuel_core_types::blockchain::primitives::DaBlockHeight;
 use std::sync::Arc;
 use tokio::sync::Mutex;
+
+mod algorithm_updater;
 
 pub type PoAService =
     fuel_core_poa::Service<TxPoolAdapter, BlockProducerAdapter, BlockImporterAdapter>;
@@ -176,24 +185,20 @@ pub fn init_sub_services(
     #[cfg(not(feature = "p2p"))]
     let p2p_adapter = P2PAdapter::new();
 
-    let updater_metadata = UpdaterMetadata::V0(V0Metadata {
-        new_exec_price: config.starting_gas_price,
-        min_exec_gas_price: config.min_gas_price,
-        exec_gas_price_change_percent: config.gas_price_change_percent,
-        l2_block_height: last_height.into(),
-        l2_block_fullness_threshold_percent: config.gas_price_threshold_percent,
-    });
     let genesis_block_height = *genesis_block.header().height();
     let settings = consensus_parameters_provider.clone();
     let block_stream = importer_adapter.events_shared_result();
-    let l2_block_source =
-        FuelL2BlockSource::new(genesis_block_height, settings, block_stream);
-    let metadata_storage = StructuredStorage::new(database.gas_price().clone());
-    let update_algo =
-        FuelGasPriceUpdater::init(updater_metadata, l2_block_source, metadata_storage)?;
-    let gas_price_service =
-        fuel_core_gas_price_service::new_service(last_height, update_algo)?;
-    let next_algo = gas_price_service.shared.clone();
+
+    let gas_price_init = algorithm_updater::InitializeTask::new(
+        config.clone(),
+        genesis_block_height,
+        settings,
+        block_stream,
+        database.gas_price().clone(),
+        database.on_chain().clone(),
+    )?;
+    let next_algo = gas_price_init.shared_data();
+    let gas_price_service = ServiceRunner::new(gas_price_init);
 
     let gas_price_provider = FuelGasPriceProvider::new(next_algo);
     let txpool = fuel_core_txpool::new_service(