+++ crates/services/executor/src/executor.rs
@@ -920,7 +920,8 @@ where
             Transaction::Script(_)
             | Transaction::Create(_)
             | Transaction::Upgrade(_)
-            | Transaction::Upload(_) => Ok(()),
+            | Transaction::Upload(_)
+            | Transaction::Blob(_) => Ok(()),
         }
     }
 
@@ -940,6 +941,7 @@ where
             }
             Transaction::Upgrade(tx) => tx.max_gas(gas_costs, fee_params),
             Transaction::Upload(tx) => tx.max_gas(gas_costs, fee_params),
+            Transaction::Blob(tx) => tx.max_gas(gas_costs, fee_params),
         };
         if actual_max_gas > claimed_max_gas {
             return Err(ForcedTransactionFailure::InsufficientMaxGas {
@@ -1006,6 +1008,15 @@ where
                 memory,
             ),
             CheckedTransaction::Upload(tx) => self.execute_chargeable_transaction(
+                tx,
+                header,
+                coinbase_contract_id,
+                gas_price,
+                execution_data,
+                storage_tx,
+                memory,
+            ),
+            CheckedTransaction::Blob(tx) => self.execute_chargeable_transaction(
                 tx,
                 header,
                 coinbase_contract_id,
@@ -1364,9 +1375,13 @@ where
             .used_gas
             .checked_add(used_gas)
             .ok_or(ExecutorError::GasOverflow)?;
-        execution_data
-            .message_ids
-            .extend(receipts.iter().filter_map(|r| r.message_id()));
+
+        if !reverted {
+            execution_data
+                .message_ids
+                .extend(receipts.iter().filter_map(|r| r.message_id()));
+        }
+
         let status = if reverted {
             TransactionExecutionResult::Failed {
                 result: Some(state),