+++ crates/types/src/lib.rs
@@ -60,6 +60,10 @@ pub mod fuel_vm {
         storage::ContractsAssetKey,
         storage::ContractsStateKey,
         storage::UploadedBytecode,
+        storage::{
+            BlobBytes,
+            BlobData,
+        },
         util,
     };
 }